Minor cleanup of all previous logger references.
Fixed error in Abaqus writer where coupling orientation needed to be in all CAPS for it to be recognized during reading inp file.

Refactored remaining fem.io components to be located in individual subfolders where typical read/write is split into separate files.

Changed the hinge2coupling method so that it adds couplings to assembly level and not on part level. Could